Removed image includes and fixed real/float conflicts

+++ src/scenes/bright_scene.hpp
@@ -8,7 +8,6 @@
 #include "camera.hpp"
 #include "colour.hpp"
 #include "hittable_list.hpp"
-#include "image.hpp"
 #include "material.hpp"
 #include "material_manager.hpp"
 #include "sphere.hpp"
@@ -19,7 +18,7 @@ auto bright_scene() {
   // Image
   const real aspect_ratio = 16.0 / 9.0;
   const int image_width = 1200;
-  const int image_height = static_cast<int>(image_width / aspect_ratio);
+  const int image_height = image_width / aspect_ratio;
 
   hittable_list world;
